ResearchProject_Code/ModularApp/requirements.txt
streamlit
pandas
numpy
scikit-learn
pandas-ta
plotly
ccxt
beautifulsoup4
google-cloud-language
requests
